chore: gitlab.com/gitlab org/api/client/.go/1.x bump

See merge request https://gitlab.com/gitlab-org/container-registry/-/merge_requests/2733

Merged-by: Hayley Swimelar <hswimelar@gitlab.com>
Approved-by: SAhmed <sahmed@gitlab.com>
Approved-by: Hayley Swimelar <hswimelar@gitlab.com>

+++ cmd/internal/flake-tracker/report/report_test.go
@@ -192,7 +192,7 @@ type fakePipelinesClient struct {
 	err    error
 }
 
-func (f *fakePipelinesClient) GetPipelineTestReport(_, _ int, _ ...gitlab.RequestOptionFunc) (*gitlab.PipelineTestReport, *gitlab.Response, error) {
+func (f *fakePipelinesClient) GetPipelineTestReport(_, _ int64, _ ...gitlab.RequestOptionFunc) (*gitlab.PipelineTestReport, *gitlab.Response, error) {
 	return f.report, &gitlab.Response{}, f.err
 }
